backend/gcp/requirements.txt
Flask==2.1.0; python_version > '3.6'
Flask==2.0.3; python_version < '3.7'
flask-cors==3.0.10
requests==2.27.1
python-dotenv==1.0.0
google-cloud-datastore==2.15.1
gunicorn==20.1.0
google-auth==2.17.1  
google-api-python-client==2.83.0       
google-cloud==0.34.0